requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
typing_extensions==4.0.1
yarl==1.7.2
pytest==6.2.5
pytest-mock==3.6.1
greenlet==1.1.2
sqlalchemy==1.4.31

